implement userCoursesValidator
